rename pixel_transforms -> pixels
