+++ src/highdicom/pixels.py
